User can empty cart

+++ app/controllers/orders_controller.rb
@@ -5,19 +5,28 @@ class OrdersController < ApplicationController
 
   def update
     @order = Order.find(params[:id])
-    order_items = params[:order][:order_items_attributes]
 
-    order_items.each do |key, value|
-      order_item = OrderItem.find(value['id'])
+    if params['update_cart']
+      order_items = params[:order][:order_items_attributes]
+      order_items.each do |key, value|
+        order_item = OrderItem.find(value['id'])
 
-      if value['quantity'].to_i == 0
-        order_item.destroy
-      else
-        order_item.update_attributes(quantity: value['quantity'].to_i)
+        if value['quantity'].to_i == 0
+          order_item.destroy
+        else
+          order_item.update_attributes(quantity: value['quantity'].to_i)
+        end
       end
-    end
+      flash[:success] = 'Your cart was updated.'
+      redirect_to @order
 
-    flash[:success] = 'Your cart was updated.'
-    redirect_to @order
-  end
+    elsif params['empty_cart']
+      @order.destroy
+      session[:order_id] = nil
+      flash[:info] = 'Your cart has been cleared out.'
+      redirect_to root_path
+    end
+  end#update
+
+  private
 end
